refactor: consolidate secrets and variables commands
- Replaced separate commands for secrets and variables with unified commands.
- Improved code structure and readability.
- Simplified command registration.
- Removed redundant code.
- Enhanced maintainability.
